+++ src/python/arcor2_scene/scripts/scene.py
@@ -548,7 +548,7 @@ def get_started() -> RespT:
     return jsonify(started)
 
 
-@app.route("/health")
+@app.route("/healthz/ready")
 def health():
     """Healthcheck.
     ---
@@ -557,10 +557,13 @@ def health():
             - Service
         description: Gets service health status.
         responses:
-            204:
+            200:
               description: The service is up and running.
+            503:
+              description: The service is up but not ready to serve requests yet.
     """
-    return Response(status=204)
+
+    return Response(status=200)
 
 
 def main() -> None: